Add new collision detection tool, the scene has been updated to support world floor height settings, and the bird launch controller has been optimized to consider bottom edge contact of objects. The collision handling logic in the scene has been refactored to use the new tool.
